+++ apps/api/src/app/endpoints/data-providers/ghostfolio/ghostfolio.module.ts
@@ -1,11 +1,13 @@
 import { RedisCacheModule } from '@ghostfolio/api/app/redis-cache/redis-cache.module';
-import { ConfigurationModule } from '@ghostfolio/api/services/configuration/configuration.module';
+import { ConfigurationService } from '@ghostfolio/api/services/configuration/configuration.service';
 import { CryptocurrencyModule } from '@ghostfolio/api/services/cryptocurrency/cryptocurrency.module';
 import { AlphaVantageService } from '@ghostfolio/api/services/data-provider/alpha-vantage/alpha-vantage.service';
 import { CoinGeckoService } from '@ghostfolio/api/services/data-provider/coingecko/coingecko.service';
+import { YahooFinanceDataEnhancerService } from '@ghostfolio/api/services/data-provider/data-enhancer/yahoo-finance/yahoo-finance.service';
+import { DataProviderModule } from '@ghostfolio/api/services/data-provider/data-provider.module';
+import { DataProviderService } from '@ghostfolio/api/services/data-provider/data-provider.service';
 import { EodHistoricalDataService } from '@ghostfolio/api/services/data-provider/eod-historical-data/eod-historical-data.service';
 import { FinancialModelingPrepService } from '@ghostfolio/api/services/data-provider/financial-modeling-prep/financial-modeling-prep.service';
-import { GhostfolioService } from '@ghostfolio/api/services/data-provider/ghostfolio/ghostfolio.service';
 import { GoogleSheetsService } from '@ghostfolio/api/services/data-provider/google-sheets/google-sheets.service';
 import { ManualService } from '@ghostfolio/api/services/data-provider/manual/manual.service';
 import { RapidApiService } from '@ghostfolio/api/services/data-provider/rapid-api/rapid-api.service';
@@ -17,15 +19,14 @@ import { SymbolProfileModule } from '@ghostfolio/api/services/symbol-profile/sym
 
 import { Module } from '@nestjs/common';
 
-import { DataEnhancerModule } from './data-enhancer/data-enhancer.module';
-import { YahooFinanceDataEnhancerService } from './data-enhancer/yahoo-finance/yahoo-finance.service';
-import { DataProviderService } from './data-provider.service';
+import { GhostfolioController } from './ghostfolio.controller';
+import { GhostfolioService } from './ghostfolio.service';
 
 @Module({
+  controllers: [GhostfolioController],
   imports: [
-    ConfigurationModule,
     CryptocurrencyModule,
-    DataEnhancerModule,
+    DataProviderModule,
     MarketDataModule,
     PrismaModule,
     PropertyModule,
@@ -35,6 +36,7 @@ import { DataProviderService } from './data-provider.service';
   providers: [
     AlphaVantageService,
     CoinGeckoService,
+    ConfigurationService,
     DataProviderService,
     EodHistoricalDataService,
     FinancialModelingPrepService,
@@ -43,13 +45,13 @@ import { DataProviderService } from './data-provider.service';
     ManualService,
     RapidApiService,
     YahooFinanceService,
+    YahooFinanceDataEnhancerService,
     {
       inject: [
         AlphaVantageService,
         CoinGeckoService,
         EodHistoricalDataService,
         FinancialModelingPrepService,
-        GhostfolioService,
         GoogleSheetsService,
         ManualService,
         RapidApiService,
@@ -61,7 +63,6 @@ import { DataProviderService } from './data-provider.service';
         coinGeckoService,
         eodHistoricalDataService,
         financialModelingPrepService,
-        ghostfolioService,
         googleSheetsService,
         manualService,
         rapidApiService,
@@ -71,15 +72,12 @@ import { DataProviderService } from './data-provider.service';
         coinGeckoService,
         eodHistoricalDataService,
         financialModelingPrepService,
-        ghostfolioService,
         googleSheetsService,
         manualService,
         rapidApiService,
         yahooFinanceService
       ]
-    },
-    YahooFinanceDataEnhancerService
-  ],
-  exports: [DataProviderService, ManualService, YahooFinanceService]
+    }
+  ]
 })
-export class DataProviderModule {}
+export class GhostfolioModule {}