+++ apps/client/src/app/components/admin-settings/ghostfolio-premium-api-dialog/ghostfolio-premium-api-dialog.component.ts
@@ -1,3 +1,5 @@
+import { DataService } from '@ghostfolio/client/services/data.service';
+import { PROPERTY_API_KEY_GHOSTFOLIO } from '@ghostfolio/common/config';
 import { GfPremiumIndicatorComponent } from '@ghostfolio/ui/premium-indicator';
 
 import { CommonModule } from '@angular/common';
@@ -30,10 +32,28 @@ import { GhostfolioPremiumApiDialogParams } from './interfaces/interfaces';
 export class GfGhostfolioPremiumApiDialogComponent {
   public constructor(
     @Inject(MAT_DIALOG_DATA) public data: GhostfolioPremiumApiDialogParams,
+    private dataService: DataService,
     public dialogRef: MatDialogRef<GfGhostfolioPremiumApiDialogComponent>
   ) {}
 
   public onCancel() {
     this.dialogRef.close();
+  }
+
+  public onSetGhostfolioApiKey() {
+    let ghostfolioApiKey = prompt(
+      $localize`Please enter your Ghostfolio API key:`
+    );
+    ghostfolioApiKey = ghostfolioApiKey?.trim();
+
+    if (ghostfolioApiKey) {
+      this.dataService
+        .putAdminSetting(PROPERTY_API_KEY_GHOSTFOLIO, {
+          value: ghostfolioApiKey
+        })
+        .subscribe(() => {
+          this.dialogRef.close();
+        });
+    }
   }
 }