+++ libs/common/src/lib/interfaces/index.ts
@@ -40,13 +40,16 @@ import type { Position } from './position.interface';
 import type { Product } from './product';
 import type { AccountBalancesResponse } from './responses/account-balances-response.interface';
 import type { BenchmarkResponse } from './responses/benchmark-response.interface';
+import type { DataProviderGhostfolioStatusResponse } from './responses/data-provider-ghostfolio-status-response.interface';
 import type { ResponseError } from './responses/errors.interface';
+import type { HistoricalResponse } from './responses/historical-response.interface';
 import type { ImportResponse } from './responses/import-response.interface';
 import type { LookupResponse } from './responses/lookup-response.interface';
 import type { OAuthResponse } from './responses/oauth-response.interface';
 import type { PortfolioHoldingsResponse } from './responses/portfolio-holdings-response.interface';
 import type { PortfolioPerformanceResponse } from './responses/portfolio-performance-response.interface';
 import type { PublicPortfolioResponse } from './responses/public-portfolio-response.interface';
+import type { QuotesResponse } from './responses/quotes-response.interface';
 import type { ScraperConfiguration } from './scraper-configuration.interface';
 import type { Statistics } from './statistics.interface';
 import type { SubscriptionOffer } from './subscription-offer.interface';
@@ -74,12 +77,14 @@ export {
   BenchmarkProperty,
   BenchmarkResponse,
   Coupon,
+  DataProviderGhostfolioStatusResponse,
   DataProviderInfo,
   EnhancedSymbolProfile,
   Export,
   Filter,
   FilterGroup,
   HistoricalDataItem,
+  HistoricalResponse,
   Holding,
   HoldingWithParents,
   ImportResponse,
@@ -105,6 +110,7 @@ export {
   Position,
   Product,
   PublicPortfolioResponse,
+  QuotesResponse,
   ResponseError,
   ScraperConfiguration,
   Statistics,