+++ libs/common/src/lib/interfaces/responses/data-provider-ghostfolio-status-response.interface.ts
@@ -1,0 +1,4 @@
+export interface DataProviderGhostfolioStatusResponse {
+  dailyRequests: number;
+  dailyRequestsMax: number;
+}